feat: modularize the code -- put mockdata into its own directory; put Card and CardList components under components/cards/ and also separate the scss file
