+++ src/App.scss
@@ -1,28 +1,3 @@
-.profiles {
-    width: 30rem;
-    margin-left: auto;
-    margin-right: auto;
-
-
-    .github-profile {
-        margin: 1rem;
-
-        img {
-            width: 75px;
-        }
-
-        .info {
-            display: inline-block;
-            margin-left: 0.75rem;
-
-            .name {
-                font-size: 1.25rem;
-                font-weight: bold;
-            }
-        }
-    }
-}
-
 form {
     border: thin solid #ddd;
     background-color: #fbfbfb;
